Fix sonarqube issues in r5 ResourceTest
